Edit update user password methods

+++ internal/domain_model/users/user_repository.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
+	"github.com/todorpopov/school-manager/internal/domain_model"
 	"github.com/todorpopov/school-manager/internal/exceptions"
 	"github.com/todorpopov/school-manager/persistence"
 	"go.uber.org/zap"
@@ -16,7 +17,7 @@ type IUserRepository interface {
 	GetUsers(ctx context.Context, tx pgx.Tx) ([]User, *exceptions.AppError)
 	GetUserByEmail(ctx context.Context, tx pgx.Tx, email string) (*User, *exceptions.AppError)
 	UpdateUser(ctx context.Context, tx pgx.Tx, updateUser *UpdateUser) (*User, *exceptions.AppError)
-	UpdateUserPassword(ctx context.Context, tx pgx.Tx, userId int32, password string) *exceptions.AppError
+	UpdateUserPassword(ctx context.Context, tx pgx.Tx, updateUserPass *UpdateUserPassword) *exceptions.AppError
 	DeleteUser(ctx context.Context, tx pgx.Tx, userId int32) *exceptions.AppError
 }
 
@@ -30,6 +31,11 @@ func NewUserRepository(db *persistence.Database, logger *zap.Logger) *UserReposi
 }
 
 func (ur *UserRepository) CreateUser(ctx context.Context, tx pgx.Tx, createUser *CreateUser) (*User, *exceptions.AppError) {
+	validationErr := ValidateCreateUser(createUser)
+	if validationErr != nil {
+		return nil, validationErr
+	}
+
 	sql := "INSERT INTO users (first_name, last_name, email, password)" +
 		"VALUES ($1, $2, $3, $4) " +
 		"RETURNING user_id, first_name, last_name, email, password;"
@@ -56,6 +62,10 @@ func (ur *UserRepository) CreateUser(ctx context.Context, tx pgx.Tx, createUser 
 }
 
 func (ur *UserRepository) GetUserById(ctx context.Context, tx pgx.Tx, userId int32) (*User, *exceptions.AppError) {
+	if msg := domain_model.ValidateId(userId); msg != "" {
+		return nil, exceptions.NewValidationError("Invalid user ID", map[string]string{"user_id": msg})
+	}
+
 	sql := "SELECT user_id, first_name, last_name, email, password FROM users WHERE user_id = $1;"
 
 	var user User
@@ -78,6 +88,10 @@ func (ur *UserRepository) GetUserById(ctx context.Context, tx pgx.Tx, userId int
 }
 
 func (ur *UserRepository) GetUserByEmail(ctx context.Context, tx pgx.Tx, email string) (*User, *exceptions.AppError) {
+	if msg := domain_model.ValidateEmail(&email, true); msg != "" {
+		return nil, exceptions.NewValidationError("Invalid email", map[string]string{"email": msg})
+	}
+
 	sql := "SELECT user_id, first_name, last_name, email, password FROM users WHERE email = $1;"
 
 	var user User
@@ -130,6 +144,10 @@ func (ur *UserRepository) GetUsers(ctx context.Context, tx pgx.Tx) ([]User, *exc
 }
 
 func (ur *UserRepository) UpdateUser(ctx context.Context, tx pgx.Tx, updateUser *UpdateUser) (*User, *exceptions.AppError) {
+	if validationErr := ValidateUpdateUser(updateUser); validationErr != nil {
+		return nil, validationErr
+	}
+
 	sql := "UPDATE users " +
 		" SET first_name = $1, last_name = $2, email = $3 " +
 		"WHERE user_id = $4 " +
@@ -156,17 +174,21 @@ func (ur *UserRepository) UpdateUser(ctx context.Context, tx pgx.Tx, updateUser 
 	return &user, nil
 }
 
-func (ur *UserRepository) UpdateUserPassword(ctx context.Context, tx pgx.Tx, userId int32, password string) *exceptions.AppError {
+func (ur *UserRepository) UpdateUserPassword(ctx context.Context, tx pgx.Tx, updateUserPass *UpdateUserPassword) *exceptions.AppError {
+	if err := ValidateUpdateUserPassword(updateUserPass); err != nil {
+		return err
+	}
+
 	sql := "UPDATE users SET password = $1 WHERE user_id = $2;"
 
 	var err error
 	var cmdTag pgconn.CommandTag
 	if tx != nil {
 		ur.logger.Debug("Updating user password in transaction")
-		cmdTag, err = tx.Exec(ctx, sql, password, userId)
+		cmdTag, err = tx.Exec(ctx, sql, updateUserPass.Password, updateUserPass.UserId)
 	} else {
 		ur.logger.Debug("Updating user password without transaction")
-		cmdTag, err = ur.db.Pool.Exec(ctx, sql, password, userId)
+		cmdTag, err = ur.db.Pool.Exec(ctx, sql, updateUserPass.Password, updateUserPass.UserId)
 	}
 	if err != nil {
 		ur.logger.Error("Failed to update user password", zap.Error(err))
@@ -174,7 +196,7 @@ func (ur *UserRepository) UpdateUserPassword(ctx context.Context, tx pgx.Tx, use
 	}
 
 	if cmdTag.RowsAffected() == 0 {
-		ur.logger.Error("Failed to update user password - user not found", zap.Int32("user_id", userId))
+		ur.logger.Error("Failed to update user password - user not found", zap.Int32("user_id", updateUserPass.UserId))
 		return exceptions.NewNotFoundError("User not found")
 	}
 
@@ -182,6 +204,10 @@ func (ur *UserRepository) UpdateUserPassword(ctx context.Context, tx pgx.Tx, use
 }
 
 func (ur *UserRepository) DeleteUser(ctx context.Context, tx pgx.Tx, userId int32) *exceptions.AppError {
+	if msg := domain_model.ValidateId(userId); msg != "" {
+		return exceptions.NewValidationError("Invalid user ID", map[string]string{"user_id": msg})
+	}
+
 	sql := "DELETE FROM users WHERE user_id = $1;"
 
 	var err error
